Use `dkms` bin from dkms-cli

+++ scripts/test_sad.sh
@@ -1,5 +1,10 @@
-alias dkms-dev-cli="./target/release/dkms-dev-cli"
+dkms="./dkms"
+
+if [ ! -e "$dkms" ]; then
+    echo "$dkms bin not found. Please see README"
+    exit 1
+fi
 
 echo -n '{"hello":"world","d":""}' > /tmp/said.json
 
-dkms-dev-cli said sad -f /tmp/said.json
+$dkms said sad -f /tmp/said.json
